+++ extension/chrome/elements/pgp_block.ts
@@ -10,18 +10,18 @@ import { Ui } from '../../js/common/browser.js';
 import { BrowserMsg } from '../../js/common/extension.js';
 import { Lang } from '../../js/common/lang.js';
 import { Api, AuthError } from '../../js/common/api/api.js';
-import { VerifyRes, DecryptErrTypes, FormatError, PgpMsg, Pgp } from '../../js/common/core/pgp.js';
+import { VerifyRes, DecryptErrTypes, FormatError, PgpMsg } from '../../js/common/core/pgp.js';
 import { Mime, MsgBlock } from '../../js/common/core/mime.js';
 import { Google, GmailResponseFormat } from '../../js/common/api/google.js';
 import { Buf } from '../../js/common/core/buf.js';
 import { BackendRes, Backend } from '../../js/common/api/backend.js';
 import { Assert } from '../../js/common/assert.js';
 import { Xss } from '../../js/common/platform/xss.js';
-import { Keyserver } from '../../js/common/api/keyserver.js';
 import { Settings } from '../../js/common/settings.js';
 import { Url } from '../../js/common/core/common.js';
 import { View } from '../../js/common/view.js';
 import { PgpBlockViewAttachmentsModule } from './pgp_block_attachmens_module.js';
+import { PgpBlockViewSignatureModule } from './pgp_block_signature_module.js';
 
 export class PgpBlockView extends View { // tslint:disable-line:variable-name
 
@@ -31,7 +31,7 @@ export class PgpBlockView extends View { // tslint:disable-line:variable-name
   private readonly hasChallengePassword: boolean;
   private readonly isOutgoing: boolean;
   private readonly short: string | undefined;
-  private readonly senderEmail: string | undefined;
+  public readonly senderEmail: string | undefined;
   private readonly msgId: string | undefined;
   private readonly encryptedMsgUrlParam: Buf | undefined;
 
@@ -44,9 +44,10 @@ export class PgpBlockView extends View { // tslint:disable-line:variable-name
   private passwordMsgLinkRes: BackendRes.FcLinkMsg | undefined;
   private adminCodes: string[] | undefined;
   private userEnteredMsgPassword: string | undefined;
-  private doNotSetStateAsReadyYet = false;
+  public doNotSetStateAsReadyYet = false;
 
   public readonly attachmentsModule: PgpBlockViewAttachmentsModule;
+  public readonly signatureModule: PgpBlockViewSignatureModule;
 
   constructor() {
     super();
@@ -63,7 +64,9 @@ export class PgpBlockView extends View { // tslint:disable-line:variable-name
     this.msgId = Assert.urlParamRequire.optionalString(uncheckedUrlParams, 'msgId');
     this.encryptedMsgUrlParam = uncheckedUrlParams.message ? Buf.fromUtfStr(Assert.urlParamRequire.string(uncheckedUrlParams, 'message')) : undefined;
     this.signature = uncheckedUrlParams.signature === true ? true : (uncheckedUrlParams.signature ? String(uncheckedUrlParams.signature) : undefined);
+    // modules
     this.attachmentsModule = new PgpBlockViewAttachmentsModule(this);
+    this.signatureModule = new PgpBlockViewSignatureModule(this);
   }
 
   async render() {
@@ -155,7 +158,7 @@ export class PgpBlockView extends View { // tslint:disable-line:variable-name
     return `<div class="button long ${addClasses}" style="margin:30px 0;" target="cryptup">${text}</div>`;
   }
 
-  private setFrameColor(color: 'red' | 'green' | 'gray') {
+  public setFrameColor(color: 'red' | 'green' | 'gray') {
     if (color === 'red') {
       $('#pgp_background').removeClass('pgp_secure').removeClass('pgp_neutral').addClass('pgp_insecure');
     } else if (color === 'green') {
@@ -198,87 +201,6 @@ export class PgpBlockView extends View { // tslint:disable-line:variable-name
       return hashed;
     }
     return pwd;
-  }
-
-  private async renderPgpSignatureCheckMissingPubkeyOptions(signerLongid: string, senderEmail: string | undefined): Promise<void> { // don't have appropriate pubkey by longid in contacts
-    const render = (note: string, action: () => void) => $('#pgp_signature').addClass('neutral').find('.result').text(note).click(this.setHandler(action));
-    try {
-      if (senderEmail) { // we know who sent it
-        const [senderContactByEmail] = await Store.dbContactGet(undefined, [senderEmail]);
-        if (senderContactByEmail && senderContactByEmail.pubkey) {
-          render(`Fetched the right pubkey ${signerLongid} from keyserver, but will not use it because you have conflicting pubkey ${senderContactByEmail.longid} loaded.`, () => undefined);
-          return;
-        } // ---> and user doesn't have pubkey for that email addr
-        const { pubkey, pgpClient } = await Keyserver.lookupEmail(this.acctEmail, senderEmail);
-        if (!pubkey) {
-          render(`Missing pubkey ${signerLongid}`, () => undefined);
-          return;
-        } // ---> and pubkey found on keyserver by sender email
-        const { keys: [keyDetails] } = await BrowserMsg.send.bg.await.pgpKeyDetails({ pubkey });
-        if (!keyDetails || !keyDetails.ids.map(ids => ids.longid).includes(signerLongid)) {
-          render(`Fetched sender's pubkey ${keyDetails.ids[0].longid} but message was signed with a different key: ${signerLongid}, will not verify.`, () => undefined);
-          return;
-        } // ---> and longid it matches signature
-        await Store.dbContactSave(undefined, await Store.dbContactObj({
-          email: senderEmail, pubkey, client: pgpClient, expiresOn: await Pgp.key.dateBeforeExpiration(pubkey)
-        })); // <= TOFU auto-import
-        render('Fetched pubkey, click to verify', () => window.location.reload());
-      } else { // don't know who sent it
-        const { pubkey, pgpClient } = await Keyserver.lookupLongid(this.acctEmail, signerLongid);
-        if (!pubkey) { // but can find matching pubkey by longid on keyserver
-          render(`Could not find sender's pubkey anywhere: ${signerLongid}`, () => undefined);
-          return;
-        }
-        const { keys: [keyDetails] } = await BrowserMsg.send.bg.await.pgpKeyDetails({ pubkey });
-        const pubkeyEmail = Str.parseEmail(keyDetails.users[0] || '').email!;
-        if (!pubkeyEmail) {
-          render(`Fetched matching pubkey ${signerLongid} but no valid email address is listed in it.`, () => undefined);
-          return;
-        }
-        const [conflictingContact] = await Store.dbContactGet(undefined, [pubkeyEmail]);
-        if (conflictingContact && conflictingContact.pubkey) {
-          render(`Fetched matching pubkey ${signerLongid} but conflicting key is in local contacts ${conflictingContact.longid} for email ${pubkeyEmail}, cannot verify.`, () => undefined);
-          return;
-        }
-        render(`Fetched matching pubkey ${signerLongid}. Click to load and use it.`, async () => {
-          await Store.dbContactSave(undefined, await Store.dbContactObj({
-            email: pubkeyEmail, pubkey, client: pgpClient, expiresOn: await Pgp.key.dateBeforeExpiration(pubkey)
-          })); // TOFU manual import
-          window.location.reload();
-        });
-      }
-    } catch (e) {
-      if (Api.err.isSignificant(e)) {
-        Catch.reportErr(e);
-        render(`Could not load sender pubkey ${signerLongid} due to an error.`, () => undefined);
-      } else {
-        render(`Could not look up sender's pubkey due to network error, click to retry.`, () => window.location.reload());
-      }
-    }
-  }
-
-  private renderPgpSignatureCheckResult(signature: VerifyRes | undefined) {
-    if (signature) {
-      const signerEmail = signature.contact ? signature.contact.name || this.senderEmail : this.senderEmail;
-      $('#pgp_signature > .cursive > span').text(signerEmail || 'Unknown Signer');
-      if (signature.signer && !signature.contact) {
-        this.doNotSetStateAsReadyYet = true; // so that body state is not marked as ready too soon - automated tests need to know when to check results
-        this.renderPgpSignatureCheckMissingPubkeyOptions(signature.signer, this.senderEmail).then(() => { // async so that it doesn't block rendering
-          this.doNotSetStateAsReadyYet = false;
-          Ui.setTestState('ready');
-          $('#pgp_block').css('min-height', '100px'); // signature fail can have a lot of text in it to render
-          this.resizePgpBlockFrame();
-        }).catch(Catch.reportErr);
-      } else if (signature.match && signature.signer && signature.contact) {
-        $('#pgp_signature').addClass('good');
-        $('#pgp_signature > .result').text('matching signature');
-      } else {
-        $('#pgp_signature').addClass('bad');
-        $('#pgp_signature > .result').text('signature does not match');
-        this.setFrameColor('red');
-      }
-      $('#pgp_signature').css('block');
-    }
   }
 
   private renderFutureExpiration(date: string) {
@@ -347,7 +269,7 @@ export class PgpBlockView extends View { // tslint:disable-line:variable-name
 
   private async decideDecryptedContentFormattingAndRender(decryptedBytes: Buf, isEncrypted: boolean, sigResult: VerifyRes | undefined, plainSubject?: string) {
     this.setFrameColor(isEncrypted ? 'green' : 'gray');
-    this.renderPgpSignatureCheckResult(sigResult);
+    this.signatureModule.renderPgpSignatureCheckResult(sigResult);
     const publicKeys: string[] = [];
     let renderableAtts: Att[] = [];
     let decryptedContent = decryptedBytes.toUtfStr();
